Initial Commit with Functional VGA Generator
This is the first commit that includes a functional VGA generator for
AVR microcontrollers. This project currently targets the ATmega1284p
specifically.

This is merely a static frame buffer. The next step is to implement a
command set so that another microcontroller may set pixels and perform
other graphics related operations.